platform.Memd: directly trigger OOM dbus signal
Until now, the platform.Memd test was able to forge kernel messages
because the anomaly_detector wasn't strict with it's input. With
journald, kernel messages aren't forgable, so it needs to directly send
the OOM-kill notification itself instead of forging an OOM message in the
logs that would be picked up by the anomaly_detector.

BUG=chromium:874698
TEST=tast run $DUT platform.Memd

Change-Id: I7be17ef2a11f89bb5d33dc1ab212fea36c02efb5