Update patch set 2
Patch Set 2:

The algorithm in tests should be changed to sha256 but it doesn't affect the test result. Because platform.DMVerity tests don't verifies the algorithm or test the rootfs. These tests create a logical device and verifies verity mechanism with the algorithm stated in the test script. But, since dm_verity algorithm is changed on the device, changing the algorithm on the tests also needed.

Patch-set: 2
Reviewer: Gerrit User 1344767 <1344767@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>