Fix flakiness of ownership tests
Fixes the flakiness of session.OwnershipAPI by restarting Chrome with
--profile-requires-policy=true. This prevents a crash on startup since
the profile is created synchronously, but Chrome didn't know whether to
expect policy.

The runtime of the unfixed test ranges from 2.5 seconds to 5.7s (on
Chell). This  large range might hint at the flakiness. The runtime of
the fixed test is at a pretty consistent 5.7s.

Since some other flaky ownership tests also touch policy, it seems like
a good thing to apply this fix to them as well, although it turned out
difficult to reproduce the flakiness.

BUG=chromium:950812
TEST=Check https://tastboard.corp.google.com/test for
       session.GuestAndActualSession
       session.MultiUserPolicy
       session.OwnershipAPI
       session.OwnershipNotRetaken
       session.OwnershipRetaken
       session.OwnershipTaken
       session.RejectDuplicate
       session.RemoteOwnership
       session.RetrieveActiveSessions
       session.UserPolicyKeys

Change-Id: I20a7810e3b366321723705a9b6e04b783fa58d54