tast-tests: Fix arc.AccessibilityEvent to appropriately catch diffs
Previously, arc.AccessibilityEvent test does not handle the diffs of
event output, which results in the test passing even when there is a
difference between the expected output and the obtained one.

This CL updates how to handle outout logs.

Also, this CL abandons writing out the difference of output log to file
and now it is dumped in the error message.

BUG=none
TEST=tast run DUT_IP arc.AccessibilityEvent
Change-Id: I45972adc7dd162947609aa27b5bd9db2c66dd3b9