tast-test: Fix arc.AccessibilityEvent to appropriately catch diffs
Previously, arc.AccessibilityEvent test does not handle the diffs of
event output, which results in passing the test even when there is a
difference between the expected output and the obtained one.

This CL fixes it by appropriately returning the error.

This CL also fixes the expected output of event log.

BUG=none
TEST=tast run DUT_IP arc.AccessibilityEvent
Change-Id: I45972adc7dd162947609aa27b5bd9db2c66dd3b9