Files App Smoke Test
Ported the desktopui_FilesApp test from autotest to tast.
Added TODO for porting the image preview version of the test.
Made a cleaner abstraction of the Files App ui.
Added the ability to wait for expressions with a timeout to chrome.Conn
This is used to impose short timeouts onto waiting for UI before
printing debug info.

BUG=chromium:987755
TEST=Locally on eve and remotely on candy

Change-Id: I693a404cbd573305a173dc7f721aae00549faf57