Update patch set 5
Patch Set 5:

This still shows the same problems. I updated https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1716364 to use this. On my eve test device, that test still fails with 

2019-07-24T19:02:59.381265+00:00 INFO tast[4038]: ui.ChromeSetConsent: Waiting for org.chromium.SessionManager D-Bus service
2019-07-24T19:02:59.383504+00:00 ERR session_manager[4282]: [ERROR:policy_key.cc(68)] Policy key /var/lib/whitelist/owner.key is corrupted!
2019-07-24T19:02:59.383593+00:00 ERR session_manager[4282]: [ERROR:device_policy_service.cc(267)] Failed to load device policy key from disk.
2019-07-24T19:02:59.384094+00:00 WARNING session_manager[4282]: [WARNING:device_policy_service.cc(275)] Recovering missing owner key from policy blob!
2019-07-24T19:02:59.398673+00:00 INFO tast[4038]: ui.ChromeSetConsent: Asking session_manager to enable Chrome testing
2019-07-24T19:02:59.403529+00:00 ERR session_manager[4282]: [ERROR:policy_key.cc(102)] Tried to overwrite owner key!
2019-07-24T19:02:59.403599+00:00 ERR session_manager[4282]: [ERROR:policy_service.cc(265)] Failed to persist policy key to disk.

if I'm not logged in.

If I am logged in, it crashes the device back to the "OS Verification is OFF" boot up screen if I am (and also erases the hard disk, just for good measure).

This is on a clean eve that I just re-flashed from USB

Patch-set: 5