tast: port network_TwoShills to network/two_shills.go.
The test will ensure that a second shill process cannot be spawned.
The motivation behind the original design of the test was that there was
an issue (listed in the bug below) where DBUS would by default issue a
_queue_request for a name when an owner already existed.
This test is being ported for the purpose of migrating the
autotest network_nightly suite to Tast.

TEST=two_shills.go , manual testing on octopus DUT for false positives.
BUG=chromium:356874

Change-Id: I56748a2fdf8283f97591f6f5b886a818c836b189
Reviewed-on: https://chromium-review.googlesource.com/1717664
Tested-by: Billy Zhao <billyzhao@google.com>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>