Update patch set 1
Patch Set 1:

What's the goal of this change?
The responsibility of the ResetState is, to reuse Chrome via Precondition, tearing down any in-use Chrome state in previous tests to move back as if the chrome is in just after logging in, i.e. as same as just after chrome.New().
In the sense calling ResetState looks weird and redundant, because it is expected to be no-op at the end of chrome.New().

Patch-set: 1
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>