test-tests: default_profile: hold recover_duts lock
recover_duts's check_ethernet notices that we 'stop shill', which
results in a brief loss of connectivity. The subsequent removal / reset
of USB ethernet devices is not desireable -- and if enough factors
collide, it may even cause an unexpected reboot.

Grab the lock instead.

Move this back to informational for a little while while we're at it.

BUG=none
TEST=run it

Change-Id: Ie8af0394e0e76c03b94a334bcc23ef1c08c8846a