tast: Add arc.ImeBlockingVK
arc.ImeBlockingVK tests that the virtual keyboard is properly hidden
when a dialog with ALT_FOCUSABLE_IM is opened. In stock Android,
the VK should be shown behind the dialog's dim surface in such a case.
As ARC++ does not support showing CrOS VK behind an Android dialog,
we should hide the VK itself here.

TEST=tast run eve-DUT arc.ImeBlockingVK
BUG=b:136527424

Change-Id: I7f1e8fb34c76f989008a90f1a212494bd533d66b