platform.MemoryPressure: don't log times for discarded tabs
I noticed we were getting some very low tab switch numbers
occassionally for the cold tab switches and determined this was
because we were trying to switch to discarded tabs but still reporting
a non-zero switch time.  Having some discarded tabs in that phase of
the test is expected, but we shouldn't report it as a successful
switch.  Instead, let's report whether a tab activiation was
successful or not and if not 0 for either a discarded tab or any other
error.

Also fix a lint error for the comment for googleLogIn().

BUG=none
TEST=run platform.MemoryPressure

Change-Id: I09395df1986a8335b7406e8b2892f117591f079d