tast: port network_UdevRename to network/udev_rename.go.
This test verifies system stability when udev is restarted or when
network interface drivers are reloaded. The interfaces should be able to
recover properly throughout the restart lifecycle.

network_UdevRename is part of the network_nightly suite and is part of
the ongoing effort to port the suite to tast.

I will cleanup the interface retrieval in a cleanup CL. (chromium:988894)
TEST=udev_rename.go, manual testing on atlas DUT.
BUG=None

Change-Id: Ic098be9f9c1e59580ec6b11ad9ffb7b5924d1620