tast: port network_UdevRename to network/udev_rename.go.
This test verifies system stability when udev is restarted or when
WiFi driver is reloaded. Network interfaces should be the same after
udev restart and WiFi driver restart.

network_UdevRename is part of the network_nightly suite and is part of
the ongoing effort to port the suite to tast.

BUG=chromium:871645
TEST=udev_rename.go, manual testing on rammus DUT.

Change-Id: Ic098be9f9c1e59580ec6b11ad9ffb7b5924d1620
Signed-off-by: Yen-lin Lai <yenlinlai@google.com>