Introduce power.ChargeScheduling
Add tests for the Peak Shift policy on Wilco devices, which controls
how much power the device draws from AC and whether or not it charges
the battery. Specifically, it can enable 3 modes:
-Use AC and charge the battery
-Use AC, but don't charge the battery
-Don't use AC, and run off just teh battery. See
https://www.chromium.org/administrators/policy-list-3#DevicePowerPeakShiftEnabled
for more information.

Currently it is important that the batteries are less than full so that
we can actually test if the DUT is charging. Therefore, there's a bit
of a hack where we try to drain the batteries as fast as possible.
This could take up to five minutes. I'm looking into the possibility
of using a battery emulator as well, but that would require some work
from the lab and some extra code.

Also, the EC is currently really slow at updating its behavior after
a policy change. I've filed a bug to try to fix this at b/138612166.
That should speed up the tests by a lot.

BUG=b:138612166,b:134153286
TEST=tast run ${DUT} power.ChargeScheduling

Change-Id: Icd33877ae71f2fbe9fc7f8c7a40425db0ddf528e