tast: Update crash tests to use real consent system
Previously, the Chrome crash tests used metrics_client -C to set
consent. However, this only touches the legacy 'Consent To Send Stats'
file; it doesn't update the policy. The legacy file is overwritten by
the policy in various cases, so we want to actually update the policy
in a more realistic manner.

The Udev crash test directly wrote premade policy files and owner files
to the disk which were not signed correctly.

For both, update the tests to use the actual Session_manager hooks to
properly set a policy of enabling metrics collection.

BUG=chromium:984807
TEST=Ran tast tests on eve and kevin

Cq-Depend: chromium:1717670, chromium:1730245
Change-Id: If41ff13835c2b89c8adffd3d8fcea4849de18d6f