Wait until Chrome animation has finished in arc.PIP
Currently, we just check the idle property of the PIP activity to
check if resizing has been done completely, but Chrome animation is
applied for PIP windows, we need to query into Ash and ensure
Chrome animation has finished and the window has an expected bounds.

Also, this CL removes some hacks in the test as the original bugs
have been already resolved.

BUG=b:131248000
BUG=b:135956156
BUG=b:135160409
TEST=tast -verbose dut my_crbook arc.PIP
Change-Id: If2cdc19f9c0be6aa905b7bddbb7b37c0bfc0e868