Update patch set 3
Patch Set 3:

As commented yesterday, I think https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1729845 is a way to go. It is already in CR+2'ed and in CQ, so I expect it will be checked in before this CL so this is no longer even temporary solution. Right?

> Patch Set 3:
> 
> hidehiko@ RS plz

Tast owner review is actually detailed code review process rather than just RS,
and we're thinking this is very important process at the current stage.
It is because Go is new for most of Chrome OS engs, and it sometimes has unique culture even compared to other Chrome OS common languages.
Also, Tast is working but still new and many features are under development.
We're trying to spread the culture/designs/approaches/directions etc. by sharing what we know via code review so that we'll be on the same page eventually.

Lessons we learned from Autotest is that no coordination will make just a chaotic situation.
Tast team is very concerning about it, and very carefully trying to avoid let Tast move into such a situation.
The current code review may take a bit extra time because of timezones,
but meantime, I really appreciate your understand, respect and cooperation to the review process for above reasons.
Note that we are working on to increase the the number of owners and timezone coverages,
so the overhead is expected to be reduced sometime soon.
Also, we may revisit here later, when Go and Tast become more popular in Chrome OS.

Best regards,
- hidehiko

Patch-set: 3
CC: Gerrit User 1337146 <1337146@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1170239 <1170239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1000313 <1000313@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>