tast: Add syslog.Watcher
Add a set of functions useful for searching for certain messages in the
system logs. Handles things like ignoring messages that were in the logs
from previous tests, as well as logs rotating during the test.

This will be used to determine the success of the CrashLoopTest. In test
mode, crash_sender will print a particular message to /var/log/messages,
and the test will look for that message to determine test success.

BUG=chromium:857317
TEST=Unit tests, ran ui.ChromeCrashLoop using this functionality

Change-Id: I1111742e76149ae1f16432fde83318715f743af6