Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> As commented yesterday, I think https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1729845 is a way to go. It is already in CR+2'ed and in CQ, so I expect it will be checked in before this CL so this is no longer even temporary solution. Right?

Hidehiko, thank you for your review, but you did not respond to my question. Let me state it again: It is good that the current global default flags will agree with the flags that this particular test requires. But how do we document the needs of this test (or other tests) in case that the global default flags change again in the future?

Patch-set: 3
Reviewer: Gerrit User 1000221 <1000221@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>