video.DecodeAccelPerf: Fix capped video decoder performance test.
CL crrev.com/c/1718086 migrated tests from bintest to gtest. An error in the
gtest filter for the capped performance tests stopped the test from running.

The test is run with gtest_repeat=-1 to run it in a loop while measuring CPU
usage. Since the incorrect filter causes no testcases to be run, an endless
stream of data is printed to the command line. CPU usage is measured for 20s, in
that time a 877MB log file is created.

TEST=video.DecodeAccelPerfH2641080P60FPS on eve

BUG=b:138822793

Change-Id: I866b306a72c3b98cace6b98fa901c29b3a694744