camera.HAL3StillCapture: Lengthen the timeout.
It takes 1m34s on my Nocturne locally with the normal office lighting
condition, but 2m3s in the lab.

Lengthen the test timeout to 5 minutes to have a room for sku
difference (i7 vs celeron) and lighting difference (lab is dark and may
make the exposure time longer / 3A convergence slower).

BUG=b:138730940
TEST=$ tast run <nocturne_in_lab> camera.HAL3StillCapture and it passed
       in 2m3s.

Change-Id: I545a5467637113c3f41980e1a41e7afbff5fe63a