Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > As commented yesterday, I think https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1729845 is a way to go. It is already in CR+2'ed and in CQ, so I expect it will be checked in before this CL so this is no longer even temporary solution. Right?
> 
> Hidehiko, thank you for your review, but you did not respond to my question. Let me state it again: It is good that the current global default flags will agree with the flags that this particular test requires. But how do we document the needs of this test (or other tests) in case that the global default flags change again in the future?

Hi Ilja, sorry that I didn't reply to your question, and thank you for ping.

> It is good that the current global default flags will agree with the flags that this particular test requires.

I think no-tabs at first is the expectation from not only this test but also all tests using Chrome instance in Tast. It is indirectly written here.

https://chromium.googlesource.com/chromiumos/platform/tast-tests/+/b1b3ce8b4e073eef058bbcb0f0b7fb34197e9e1f/src/chromiumos/tast/local/chrome/chrome.go#383

=== quote ===
"by closing all pages"
=== quote end ===

IMO, it's better to be commented in chrome.New and chrome.LoggedIn, too, to be clearer.
If you agree, I'm happy to make such a change.

> But how do we document the needs of this test (or other tests) in case that the global default flags change again in the future?

In this case, I think the document of chrome.New / chrome.LoggedIn helps to people find it.
Also, if you'd like, you can add line comment for chrome.New() invocation in this test, like;

// ...
// At logged in, no tab should be opened.
c := chrome.New(...)

In this case, I think the expected behavior will unlikely be changed (or if so unexpectedly, it is considered as an issue, IMO).

BTW, although this topic is interesting to be discussed, I think,
it is slightly shifted to out of the main focus of this change itself.
Also, I don't think gerrit is a good tool for the discussion.
If you think it's better to continue the discussion, could you move to other tools, like a bug or emails?

Thanks,
- hidehiko

Patch-set: 3
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>