Update patch set 1
Patch Set 1:

I know there's been talk about switching Chrome OS logging from syslog to journald (and we actually have journald logging in there already)....so it's probably a better idea to base this off the journal rather than /var/log/messages.  The autotest code actually does that as well (it tries the journal, if that fails then it reverts to /var/log/messages...so if you want to leave that in as a fallback that's fine).

Patch-set: 1