tast-tests: crostini: Moved startup test to crostini package
Unlike the other refactorings, we will not use the artifact precondition
(yet) for this test. There are several changes to how startup will be
done in the pipeline so it is better not to create conflicts yet.

This refactoring just changes the package/name of the test and adopts
the chrome precondition.

BUG=chromium:980048
TEST=tast -verbose run localhost:33333 "crostini.StartupPerf"

Change-Id: I78969709f25c1365c87641445c22cbe7ef8e09b5