tast: Remove vkb.TapKey workaround.
As the original issue on Android side was fixed on both N and P,
we no longer need the hack in vkb.TapKey. As the hack introduced
flakiness, we should not rely on it and instead use the textfield
side attribute if autocapitalization is unnecessary.

For the background, see https://crrev.com/c/1580840

TEST=arc.ChromeVirtualKeyboard
BUG=b:137063310

Change-Id: Ifcf11b5ae3a5ed0fc76e97af81b9a02992574150
Reviewed-on: https://chromium-review.googlesource.com/1743889
Tested-by: Tetsui Ohkubo <tetsui@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>