Update patch set 14
Patch Set 14:

re: The discussion on whether it makes sense to run this command if battery is not present:

We have built consensus around the idea that it would make sense to run this command even if the device doesn't have a battery. In this case, we could validate that bad/incorrect values aren't being reported.

Patch-set: 14
Reviewer: Gerrit User 1301140 <1301140@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>