Introduce power.WilcoPeakShift
The Peak Shift policy on Wilco devices uses the EC to schedule when the
DUT uses AC power. Specifically, by using the policy it is possible to
schedule 3 different modes:
 - Use AC and charge the battery (max AC usage)
 - Use AC, but don't charge the battery (medium AC usage)
 - Don't use AC, and run off just the battery. (no AC usage)
For more information, see
https://www.chromium.org/administrators/policy-list-3#DevicePowerPeakShiftEnabled

In order to test all aspects of the policy, it must be possible for the
EC to charge the battery. That would require the battery level to be
less than 100%, but since in the lab all devices are always plugged into
AC, the batteries are kept at 100%. In order to test the charging
behavior this test would need to drain the batteries somewhat, which
could take ~10 minutes. Therefore, this test only verifies that the
device can toggle AC usage on and off, and doesn't test for battery
charging/no-charging.

TODO(b/138940522): Add a more thorough, slower, manual test for
all aspects of the policy.

Also, the EC currently takes a minute to update its behavior after a
policy change. This mean this test has a timeout of 4 minutes, but from
test runs it usually takes about 2.5 minutes. I've filed a bug to try to
fix this at b/138612166.

BUG=b:138612166,b:134153286,b:138938829
TEST=tast run ${DUT} power.WilcoPeakShift

Change-Id: Icd33877ae71f2fbe9fc7f8c7a40425db0ddf528e