tast-tests: ensure hyperthreading can be enabled in ARC
For devices with hyperthreading capable CPUs, the user can change
chrome://flags to enable hyperthreading even though it is currently
disabled by default.

Ensure that the ARC container is able to use the extra logical cores
when they come online after login.

Tests require that we login again with hyperthreading and ARC enabled so
we cannot reuse the arc.Booted() precondition.

NB: This test is currently failing, and would have since ~R74.

BUG=b:139190638
TEST=tast run {IP} arc.ContainerMount arc.MiniContainer \
     arc.CgroupsCpucontroller arc.CgroupsCpucontrollerHyperthread

Change-Id: I8a6a4b9f27d70373c7f718d10c1a5d2a33504cc3