camera.DecodeAccelJPEGPerf: Add decode latencies to metrics
This CL parses the JPEG decode latency reported from the test binary and
adds it to the metrics. It also waits for the unittest to finish instead
of killing it for the complete output logs.

BUG=b:120057531
TEST=emerge tast-local-tests-cros
TEST=Run the test and verify results-chart.json has the new metrics.

Change-Id: I2699ea113e9f36d3ce1fe951c9b2f0ca077cfe4c
Reviewed-on: https://chromium-review.googlesource.com/1745202
Tested-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>