tast-tests: crostini: Added tests for different GUI toolkits
These tests install a toolkit in the container, and verify that it
renders correctly via a screenshot. The advantage of these tests over
the VeryfyApp* tests is that we can easily modify and distribute the
code with the test.

As a side note, this test uses the new parametric testing features.

BUG=chromium:992792
TEST=tast -verbose run $DUT "crostini.Toolkit*"

Change-Id: Ic05e1192ee57917eb97723459b13b0d8c40941f8
Reviewed-on: https://chromium-review.googlesource.com/1745201
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>