Add a test for the the ChromeOS shelf.
It tests launching apps and checking that they appear via API and UI.

I've updated all uses of launchApp and isAppShown to use the same
source.

Adding a check to the TestAPI setup to wait for tast library to load for
use of tast.promisify early in a test.

BUG=chromium:991248
TEST=ui.ShelfLaunchedApps and ui.FilesAppSmoke locally,
crostini.Launcher the lab

Change-Id: I9c8e574e82df378ede9d6898c83cea08ec746c27