tast.video.DecodeAccelPerf*: Change dropped frame rate to dropped percentage.
crrev.com/c/1750534 changes the 'frame drop rate' metric to the more convenient
'frame drop percentage' in the video_decode_accelerator_perf_tests. This CL
adapts the Tast performance tests to use the new metric.

Note: Ideally this change should be submitted at the exact time of the first
Chrome uprev containing CL:1750534, but this is very hard to do. However if the
name of the metric in the json file doesn't match what the test expects, the
value is just defaulted to zero. This is acceptable while migrating to the new
metric. Additionally the reported value in most testcases is zero anyway.

TEST=video.video.DecodeAccelPerfH2641080P60FPS on eve

BUG=None

Change-Id: I94402a74f6c7a862cf9782956bca8bd9083290d9