Introduce kernel.WilcoECTelemetry
WilcoECTelemetry tests the Wilco EC's ability to respond to telemetry
commands. The Wilco EC is able to return telemetry information (such
as temperature and fan state) via sysfs: You write a binary command
to the sysfs file, the kernel driver performs some filtering on the
command to ensure it is sane, and the EC returns the binary result,
to be read from the same file. You must keep the file descriptor open
between the read and write for the response to be kept. This test
checks for end-to-end communication with the EC, and checks that the
driver performs some basic filtering and sanity checks. Since the
responses are variable and opaque binary data, it's impractical to
actually check the values of the responses.

See
https://chromium.git.corp.google.com/chromiumos/third_party/kernel/+/ea0f6a09a7a993fc7c781fd8ca675b29c42d4719/drivers/platform/chrome/wilco_ec/telemetry.c
for the kernel driver.

BUG=b:134407561
TEST=tast run ${DUT} kernel.WilcoTelemetry

Change-Id: I3954b2e13acb538155996a7ec638724e7f2ecfad