Introduce kernel.WilcoECRTC
WilcoECRTC tests the RTC contained within the EC on Wilco devices. As a
first check it reads the current time. Then, for a more detailed check,
it sets the time to a dummy time, sleeps for a bit, and reads the
time again. The RTC better have updated itself. The test attempts to
reset the RTC back to time.Now() after failure or completion.

BUG=b:139023275
TEST=tast run ${DUT} kernel.WilcoECRTC

Change-Id: I19e648ad531410d5b4d58d6be5570aed88e4c793
Reviewed-on: https://chromium-review.googlesource.com/1740391
Tested-by: Nick Crews <ncrews@chromium.org>
Commit-Ready: Nick Crews <ncrews@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>