tast-tests: new Screen rotation perf test
This CL ports the existing cheets_ScreenRotation test to Tast.
The port tries to be as faithful as possible to the original one.
But since the orignal one uses Google Inbox to grab frames, and since
Google Inbox in no longer available, it was replaced by an Activity that
is never "idle".
The other thing was changed

BUG=b:73009846
TEST=tast -verbose run $DUT arc.ScreenRotation

Change-Id: I70e300b6bf7b52f872dbd70d26cf5241bd908b58