tast-tests: new screen rotation perf test
This CL ports the existing cheets_ScreenRotation test to Tast.
http://cs/chromeos_internal/src/third_party/autotest-tests-cheets/client/site_tests/cheets_ScreenRotation/cheets_ScreenRotation.py
The port tries to be as faithful (logic-wise) as possible to the original one.

The important changes are:

1) "Google Inbox" apk was replaced, since Google Inbox is no longer available.
And older versions just don't work anymore.
The new APK is ApiDemos, and starts an activity that is in constant
motion so that frames can be grabbed.

2) The new test reports all the grabbed values, and not just the average
like the old test did.

Besides those changes, the new test uses the official Chrome API to
rotate the screen making it more robust.

BUG=b:73009846
TEST=tast -verbose run $DUT arc.ScreenRotation
     Test run Ok. Report was generated.

Change-Id: I70e300b6bf7b52f872dbd70d26cf5241bd908b58