tast-tests: new Screen rotation perf test
This CL ports the existing cheets_ScreenRotation test to Tast.
The port tries to be as faithful (logic-wise) as possible to the original one.

Although two changes were made:

1) Since the orignal test uses Google Inbox to grab frames, and since
Google Inbox in no longer available, it was replaced by an activity that
is never "idle". So, instead of capturing frames for 5 second, it
captures 50 frames. Notice that frames are being captured only if the
activity has "moving" parts. If it is idle, no new frames are captured.

2) There other change is how the performance is being reported.
Before, only one value per stat was reported. The old test took an averge
of all the values, and reported that value.
In this CL, all values from the different samples are being reported
using a Metric{Multiple:true}

Noteworthy:
- New test is more robust. Uses Chrome API to rotate screen instead of
  injecting keys, and pressing different buttons
- Main loop takes ~30 seconds (old one took ~3 minutes)

BUG=b:73009846
TEST=tast -verbose run $DUT arc.ScreenRotation
     Test run Ok. Report was generated.

Change-Id: I70e300b6bf7b52f872dbd70d26cf5241bd908b58