Introduce power.WilcoAdvancedCharging
The Advanced Charging policy on Wilco devices uses the EC to schedule
different battery charging policies. When the policy is disabled, the
device acts as normal. When the policy is enabled, then
 - If we are outside working hours, then above 90% the battery will not
   charge
 - If we are within working hours, then the battery will charge to 100%.
The policy also affects the charging rate, but it is easier just to test
for charging/no-charging. See the following link for more info:
https://www.chromium.org/administrators/policy-list-3#DeviceAdvancedBatteryChargeModeEnabled

In order to test this policy, it must be possible for the
EC to charge the battery. That would require the battery level to be
less than 100%, but since in the lab all devices are always plugged into
AC, the batteries are kept at 100%. Therefore, this test is marked as
disabled so that it must be run manually.

BUG=b:138612166,b:134153286,b:138938829
TEST=tast run ${DUT} power.WilcoAdvancedCharging

Change-Id: I803a9d901b380bc8d1241773d82c5c5e8f1054c6
Reviewed-on: https://chromium-review.googlesource.com/1747036
Tested-by: Nick Crews <ncrews@chromium.org>
Commit-Ready: Nick Crews <ncrews@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>