Introduce power.WilcoChargeModes
WilcoChargeModes tests basic control of the various charge modes that
the Wilco EC provides. Specifically, it checks that we can control
whether or not charging happens by adjusting the Charge Stop Theshold
while in Custom mode. This test is intended as an integration test and
fails to check for the various other aspects of the Charge Mode policy.

BUG=b:139201701
TEST=tast run ${DUT} power.WilcoChargeModes

Change-Id: Ie7ba3337cbe9846ee3ea0fd5bcc136c7fd1d82d8