tast-tests: crostini: Promote crostini.SanityArtifact to critical
This test was previously called vm.CrostiniStartBasic and was removed
from CQ due to flake. We have now verified that it does not flake for
sufficiently many runs (barring
https://stainless.corp.google.com/browse/chromeos-autotest-results/329172430-chromeos-test/
which we do not believe was caused by the test itself)

We would like to promote it to the critical tests.

BUG=chromium:993633
TEST=tast -verbose run $DUT "crostini.SanityArtifact"

Change-Id: I61bfbd332d09d0ed1bebffcdabc3619388c20e46