tast: unify network interface retrieval.
Added support to get the managed wireless interface in interface.go.
Previously, the wireless interface was obtained by matching a prefix to
the list of wireless interfaces. Now, the wireless interface is obtained
by looking through all interfaces obtained by "iw", and then choosing
the first interface that is in "managed" mode. This means that
interfaces with novel names will be picked up.

Relevant tests now use the helper function.

This change also fixes a parsing bug when obtaining the interface name
with iw_runner.go::ListInterfaces.

This change will fix the remaining flakiness in IWScan due to the DUTs
not having interfaces named "wlan0"

TEST=iw_scan.go, udev_rename.go, pci_rescan.go
BUG=chromium:988894, chromium:992083
Change-Id: Icd2dd7cafa097a7127c92e3faba4eb92d90f9f56