tast-tests: crostini: Added a precondition that invokes the installer
This is a temporary precondition. Once we have verified it is stable, we
will remove it and put its logic into the download/artifact/gpu
preconditions.

We need to add this logic so we can refactor the CrostiniFiles* tests
and eventually promote them to CQ.

BUG=chromium:980048
TEST=tast -verbose run $DUT "crostini.SanityInstaller"

Change-Id: I6933cef1c74e2d2b013de472bf39f8915aabebd7
Reviewed-on: https://chromium-review.googlesource.com/1753269
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>