tast-tests: Add test for crashing arc app.
This is a (fixed) re-submission of https://crrev.com/c/1764531, which was
reverted in https://crrev.com/c/1864642.

The reason the original CL broke is that https://crrev.com/c/1846411,
which was submitted concurrently, changed the signature of
WaitForCrashFiles. This CL updates to the new signature.

TEST=tast -verbose run 100.127.92.23 arc.AppCrash
BUG=chromium:983725

Change-Id: Ie7c731cf4e5aa7e255506f50f605bad48cd0653f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1865426
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Auto-Submit: Miriam Zimmerman <mutexlox@chromium.org>